H.Res. 189 Reported in House
Providing for consideration of the bill (H.R. 650) to amend the Truth in Lending Act to modify the definitions of a mortgage originator and a high-cost mortgage, providing for consideration of the bill (H.R. 685) to amend the Truth in Lending Act to improve upon the definitions provided for points and fees in connection with a mortgage transaction, and providing for the adoption of the concurrent resolution (S. Con. Res. 11) setting forth the congressional budget for the United States Government for fiscal year 2016 and setting forth the appropriate budgetary levels for fiscal years 2017 through 2025.

Version:  Reported in House (rh)
Date:     2015-04-13
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres189rh/xml/BILLS-114hres189rh.xml

Sponsored-By: S000250