H.Res. 189 Engrossed in House
Providing for consideration of the bill (H.R. 650) to amend the Truth in Lending Act to modify the definitions of a mortgage originator and a high-cost mortgage, providing for consideration of the bill (H.R. 685) to amend the Truth in Lending Act to improve upon the definitions provided for points and fees in connection with a mortgage transaction, and providing for the adoption of the concurrent resolution (S. Con. Res. 11) setting forth the congressional budget for the United States Government for fiscal year 2016 and setting forth the appropriate budgetary levels for fiscal years 2017 through 2025.

Version:  Engrossed in House (eh)
Date:     2015-04-14
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres189eh/xml/BILLS-114hres189eh.xml

Sponsored-By: S000250
Roll-Call: House 114-1-148 2015-04-14 Passed 239-183
Roll-Call: House 114-1-149 2015-04-14 Passed 237-185