H.Res. 231 Reported in House
Providing for consideration of the bill (H.R. 1732) to preserve existing rights and responsibilities with respect to waters of the United States, and for other purposes; providing for consideration of the conference report to accompany the concurrent resolution (S. Con. Res. 11) setting forth the congressional budget for the United States Government for fiscal year 2016 and setting forth the appropriate budgetary levels for fiscal years 2017 through 2025; and providing for consideration of the joint resolution (H.J. Res. 43) disapproving the action of the District of Columbia Council in approving the Reproductive Health Non-Discrimination Amendment Act of 2014.

Version:  Reported in House (rh)
Date:     2015-04-29
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres231rh/xml/BILLS-114hres231rh.xml

Sponsored-By: W000810