S.Res. 193 Introduced in Senate
A resolution celebrating the 50th anniversary of the historic Griswold v. Connecticut decision of the Supreme Court of the United States and expressing the sense of the Senate that the case was an important step forward in helping ensure that all people of the United States are able to use contraceptives to plan pregnancies and have healthier babies.

Version:  Introduced in Senate (is)
Date:     2015-06-04
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114sres193is/xml/BILLS-114sres193is.xml

Sponsored-By: B001277
Cosponsor-Count: 23