H.Res. 311 Introduced in House
Expressing the sense of the House of Representatives that Congress should confirm that money is not free speech and that corporations are not people for purposes of the First Amendment right to make campaign contributions by enacting a constitutional amendment overturning the decision of the Supreme Court in the case of Citizens United v. Federal Election Commission, and should restore the right of Congress and the States to impose limits on the amount of expenditures that may be made by candidates and others in support of elections for public office by enacting a constitutional amendment overturning the decision of the Supreme Court in the case of Buckley v. Valeo.

Version:  Introduced in House (ih)
Date:     2015-06-12
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres311ih/xml/BILLS-114hres311ih.xml

Sponsored-By: N000127