H.R. 3040 Introduced in House
To require the Secretary of Defense, in consultation with the Secretary of Veterans Affairs, to develop guidelines regarding the use by the Secretaries of the military departments and the Secretary of Veterans Affairs of unofficial sources of information to determine the eligibility of a member or former member of the Armed Forces for benefits and decorations when the member's service records are incomplete because of damage to the records, including records damaged by a 1973 fire at the National Personnel Records Center in St. Louis, Missouri.

Version:  Introduced in House (ih)
Date:     2015-07-13
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hr3040ih/xml/BILLS-114hr3040ih.xml

Sponsored-By: C001036
Cosponsor-Count: 16