H.Res. 412 Engrossed in House
Providing for consideration of the resolution (H. Res. 411) finding that the President has not complied with section 2 of the Iran Nuclear Agreement Review Act of 2015; providing for consideration of the bill (H.R. 3461) to approve the Joint Comprehensive Plan of Action, signed at Vienna on July 14, 2015, relating to the nuclear program of Iran; and providing for consideration of the bill (H.R. 3460) to suspend until January 21, 2017, the authority of the President to waive, suspend, reduce, provide relief from, or otherwise limit the application of sanctions pursuant to an agreement related to the nuclear program of Iran.

Version:  Engrossed in House (eh)
Date:     2015-09-10
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres412eh/xml/BILLS-114hres412eh.xml

Sponsored-By: S000250
Roll-Call: House 114-1-491 2015-09-10 Passed 243-186