H.Res. 448 Engrossed in House
Providing for consideration of the concurrent resolution (H. Con. Res. 79) directing the Clerk of the House of Representatives to make corrections in the enrollment of H.R. 719, and providing for consideration of the Senate amendment to the House amendment to the Senate amendment to the bill (H.R. 719) to require the Transportation Security Administration to conform to existing Federal law and regulations regarding criminal investigator positions, and for other purposes.

Version:  Engrossed in House (eh)
Date:     2015-09-30
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres448eh/xml/BILLS-114hres448eh.xml

Sponsored-By: C001053
Roll-Call: House 114-1-525 2015-09-30 Passed 239-187