H.Res. 844 Reported in House
Providing for consideration of the bill (H.R. 2357) to direct the Securities and Exchange Commission to revise Form S-3 so as to add listing and registration of a class of common equity securities on a national securities exchange as an additional basis for satisfying the requirements of General Instruction I.B.1. of such form and to remove such listing and registration as a requirement of General Instruction I.B.6. of such form, and providing for consideration of the bill (H.R. 5424) to amend the Investment Advisers Act of 1940 and to direct the Securities and Exchange Commission to amend its rules to modernize certain requirements relating to investment advisers, and for other purposes.

Version:  Reported in House (rh)
Date:     2016-09-06
Congress: 114

Source: https://www.govinfo.gov/content/pkg/BILLS-114hres844rh/xml/BILLS-114hres844rh.xml

Sponsored-By: S000250